Split up test model modules and create common get config / params function
