+++ tests/test_model_functions.py
@@ -10,7 +10,8 @@ from lcm.model_functions import (
     get_utility_and_feasibility_function,
 )
 from lcm.state_space import create_state_choice_space
-from tests.test_models.deterministic import get_model_config, utility
+from tests.test_models import get_model_config
+from tests.test_models.deterministic import utility
 
 
 def test_get_combined_constraint():
