+++ tests/test_models/__init__.py
@@ -1,0 +1,3 @@
+from .get_model import get_model_config, get_params
+
+__all__ = ["get_model_config", "get_params"]
