+++ tests/test_next_state.py
@@ -5,7 +5,7 @@ from pybaum import tree_equal
 from lcm.input_processing import process_model
 from lcm.interfaces import InternalModel
 from lcm.next_state import _get_stochastic_next_func, get_next_state_function
-from tests.test_models.deterministic import get_model_config
+from tests.test_models import get_model_config
 
 # ======================================================================================
 # Solve target
