+++ tests/test_regression_test.py
@@ -5,7 +5,7 @@ from pandas.testing import assert_frame_equal
 
 from lcm._config import TEST_DATA
 from lcm.entry_point import get_lcm_function
-from tests.test_models.deterministic import get_model_config, get_params
+from tests.test_models import get_model_config, get_params
 
 
 def test_regression_test():
